requirements.txt
# MTG Commander AI - Dependencies

# Core dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0

# LLM integrations (choose one or more)
openai>=1.0.0
anthropic>=0.25.0
# openai>=1.0.0  # Also works for OpenRouter (OpenAI-compatible API)
# langchain>=0.1.0  # Uncomment if using LangChain
# llama-index>=0.10.0  # Uncomment if using LlamaIndex

# Optional: Local LLM
# ollama-python>=0.1.0  # Uncomment for local models

# Vector database (for card similarity, future)
# qdrant-client>=1.7.0  # Uncomment if using Qdrant

# Development dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
mypy>=1.0.0
types-requests>=2.31.0.20240406

# Utilities
requests>=2.31.0
rich>=13.0.0  # For pretty CLI output
